Unconfirmed and partial transaction numbering in navigation menu
